feat(pipeline.yaml): Add the Slack notification for approving a deployment
This commit introduces a series of steps in the pipeline (`test -> build
-> deploy`) GitHub action which gathers some basic information about the
(previously-merged) PR, and sends a HTTP POST request to a known Slack
URL.

This results in a message being published to the all-devs channel on manage-and-deliver.

We want to maintain a deployment gate / manual approve step in our deploy-to-production step.

However, it is non-obvious when a deployment gets to the point where it is blocked by a manual approve.  This PR builds in some notification infrastructure for the humans who build and maintain the codebase.

We want to create a public, accessible, and push-based alert system to let developers know that their deployment is ready for testing.

We want to do this to reduce the time spent between a PR being merged to `main` and a deployment going out.

Building a custom Slack Workflow URL appears to be the best way to achieve this.

This PR relies on a Slack "Workflow" that I (TJWC) have created that takes in the following payload:

```ts
{
              link_to_approval_step: "https://www.github.com/...",
              developer_to_tag_slack_user_id: "slack_user_id",
              human_readable_pr_name: "Some PR name",
              pr_number: "65",
              github_username: "thomaswilsonxyz",
              repository: "ministryofjustice/hmpps-accredited-programmes-manage-and-deliver-api",
              timestamp: "2025-07-29T16:07+01:00"
}
```

This is sent via `curl` in a step in the pipeline.

It also relies on a mapping of GitHub usernames to Slack usernames, which is managed as an Environment Variable in this repo.

The URL where we actually send this information needs to remain private, and is stored in the `SLACK_WEBHOOK_URL` secret in the repo.

As developers come/go from the team, we'll need to update this environment variable.