APG-1314: Show the Status History of a Referral (#147)
* chore: API types

* chore: API types

# Conflicts:
#	server/@types/manageAndDeliverApi/index.d.ts

* APG-1314: View the Referral Status History of a Referral

**WHAT**

This commit introduces the functionality that allows a user to view the Status History of a Referral.

It does so in the (existing, but placeholder) "Stauts History" tab on the Referral Details screen.

**WHY**

Users need to see, and understand, the previous Statuses a Referral has been in, and who they were updated by (as well as any additional details).

* chore: Resolve git conflicts following PR

* chore: add ReferralStatusHistory type

* fix: eslint

* fix: remove unused import in rootLayout.njk

* lint

* chore: Remove various unused imports

* tests: Add statusHistoryPresenter and referralDetailsController tests