APG-1422: Specify "additional details" in Group Allocation
**WHAT**

This PR makes minor changes to feed through `additionalDetails` to the Group Allocation
step of the user journey.  These details are picked up by the API, and assigned to
the Status History Record which is created when the Referral is Allocated

**WHY**

When a Referral is Allocated to a Group the API now changes the Status of that
Referral to "scheduled" - in doing so we must provide some "additional details"
which is free-text given by the user to e.g. explain their decision

When these pages (the `addToGroup` and `addToGroupMoreDetails`) were created, the
API was not able to take that data or do anything with it.  An API change has
been proposed (on branch `APG-1422/add-details-when-assigning-to-group`) which
now expects an `additionalDetails` field on the body.

**HOW**

This is a relatively minor (but multiple-layer) changes which replaces the previously-
set `moreDetails` field with the `additionalDetails` field required by the API.

It also takes the `message` returned by the API and displays it as a Success notification
above the Allocated/Waitlist table in the `groupDetails` view.

Lastly it fixes an issue on the `addToGroup` view where the Region Name, not
the Group Code, was used as the heading above the question.