APG-1425: Add Filters to the Group Page (#271)
* APG-1425: Add Filters to the Group Page

**WHAT**

This commit makes wide-reaching changes to the /groups/started and /groups/not-started pages.

These are reliant in a set of changes made in the API which both change the URL and the return type of the BFF.

The changes can be summarised as:

- It plugs in both a list of Probation Delivery Units and Delivery Locations which are fed into a select and a set of checkboxes respectively
- Changes to the endpoint being called (there's a redundant `/region` component of it

**REFACTORING**

The `GroupDetailFilter` had been used (erenously) in the GroupList - these have now been separated out as two separate files again.

# Conflicts:
#	server/@types/manageAndDeliverApi/imported/index.d.ts

# Conflicts:
#	server/testutils/factories/groupsByRegionFactory.ts

* APG-1425: Refactor ReportingTeams and DeliveryLocations between Group List and Detail view

* APG-1425: Add the sex options to the Group List filter

* fix: Restore the docker-commpose.yaml file